Store and re-output arbitrary experimental terms of attributeTypes.
Newer OpenLDAP servers include X-ORDER in their schema, so this is
needed to keep Ldaptor working against OpenLDAP.

Earlier there have been individual special cases for some other
experimental terms, but as I believe they will just become more and
more common (as the OpenLDAP people get more and more enthusiastic
about rewriting the world in LDAP), I chose to add generic
experimental term processing instead.

Note that this is limited to attributeTypes only, and the experimental
terms are not understood in any way. Also please realize the schema
"parser" is very limited, and e.g. requires terms to be in a certain
order; it is the minimum amount of functionality needed to
interoperate with some popular LDAP servers. Some day, schema parsing
might be rewritten with pyparsing, but right now the need for that
just isn't strong enough.