Updated `test_proxybase.py` to use twisted.internet.task.Clock instead of a normal reactor.
I removed the TODO messages (not sure why they were there).
I updated the `createServer()` function in `testutil.py` to inpect the kwargs for
"proto_args".  If found, this disct is used as keyword args when creating the protocol
instance.  I used this to pass a Clock instead of the default reactor into the test
server.