Wrote tests to validate behavior when proxied server cannot be contacted.
The tests passed trial.
I ran the code through pyflakes and pep8.  I ignored the "lines too long"
warnings from pep8, but otherwise made the recommendations.  This did
result in double line-breaks around classes and line breaks around
functions inside functions, which may not be to the liking of everyone.
If this is a problem, let me know and I will update the line breaks.