BUGFIX: LDIF-tree server backend password modify operation did not commit.
The LDIF-tree LDAP server backend did not commit changes to the directory,
so if you queried the entry after a successful password change, the old
value would still be in `userPassword`.

Also, if the userPassword value does not start with the salted SHA1
prefix, it is now assumed to be stored in plaintext.  This is useful
for development.  It is also similar to the mechanism OpenLDAP uses.

It might be nice to add a mechanism where different hash schemes
for simple BINDs could be plugged into the server rather that be
hard-coded.