Changed quickstart client example to use endpoints; changed proxy examples to use updated ProxyBase API.
I am trying to keep the examples as simple as possible.  I know from experience
that the Ldaptor configuration module was a bit confusing to me, and seemed somewhat
like overkill in most cases.  Likewise the SRV DNS lookups that
`ldapconnector.LDAPConnector` performs for you are a bit much for simple LDAP
clients that just want to connect to a single host and port.

Also, I noticed that `LDAPConnector` and `LDAPClientCreator` (both from the
`ldapconnector` module) make use of a "private" implementation class and methods.
I would like to get those cleaned up a bit before I start recommending those
classes in the documentation.